Rename some elements and pass update logic to component instead of api
