+++ DashAI/back/api/api_v1/endpoints/datasets.py
@@ -315,8 +315,8 @@ async def delete_dataset(dataset_id: int, db: Session = Depends(get_db)):
 @router.patch("/{dataset_id}")
 async def update_dataset(
     dataset_id: int,
+    params: DatasetUpdateParams,
     db: Session = Depends(get_db),
-    params: DatasetUpdateParams = None,
 ):
     """Update a dataset name or task.
