+++ DashAI/front/src/components/custom/SelectTypeCell.jsx
